Fixed non-functional osmTiles button in import mode
Pretty confusing. Basically we build some controls in Overview for
seeing osmTiles, they get torn down on exit. In QA we build the same
controls, but I forgot to tear them down. So when we went back to
Overview, I was clicking those controls and nothing was happening, they
were unbound. The two controls were on top of each other and the QA
ones took precidence. Only noticed because I saw a slightly darker drop
shadow.